Move code to the Core
