+++ Sources/SkrybaMDCore/Arguments/Effects/ActionEffectable.swift
@@ -1,0 +1,3 @@
+protocol ActionEffectable {
+    func run(value: String?)
+}
