+++ Sources/SkrybaMDCore/Arguments/Effects/GenerateFileWithoutPathEffect.swift
@@ -1,0 +1,7 @@
+struct GenerateFileWithoutPathEffect: ActionEffectable {
+    private let engine = Engine.self
+    
+    func run(value: String?) {
+        engine.run(fileName: value)
+    }
+}
