+++ Sources/SkrybaMDCore/Arguments/Effects/UndefinedEffect.swift
@@ -1,0 +1,11 @@
+struct UndefinedEffect: ActionEffectable {
+    private let argument: String
+    init(argument: String) {
+        self.argument = argument
+    }
+    
+    func run(value: String?) {
+        GenerateFileWithoutPathEffect()
+            .run(value: argument)
+    }
+}
