+++ Sources/SkrybaMDCore/Globals.swift
@@ -1,0 +1,1 @@
+let productName = "SkrybaMD"
