+++ Sources/SkrybaMDCore/Utils/Array.swift
@@ -1,0 +1,9 @@
+extension Array {
+    public subscript(safe x: Int) -> Element? {
+        if self.count > x {
+            return self[x]
+        }
+        
+        return nil
+    }
+}
